separate concerns by moving structs into separate files
